Fix build and compile errors
