+++ frontend/app/layout.tsx
@@ -1,19 +1,6 @@
-"use client";
 import type { Metadata } from "next";
 import "./globals.css";
-import { ThemeProvider, createTheme } from "@mui/material/styles";
-import CssBaseline from "@mui/material/CssBaseline";
-
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: "#3b82f6",
-    },
-    secondary: {
-      main: "#f97316",
-    },
-  },
-});
+import ClientThemeProvider from "../components/ClientThemeProvider";
 
 export const metadata: Metadata = {
   title: "AEB Compositor",
@@ -24,10 +11,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
   return (
     <html lang="en">
       <body className="antialiased">
-        <ThemeProvider theme={theme}>
-          <CssBaseline />
-          {children}
-        </ThemeProvider>
+        <ClientThemeProvider>{children}</ClientThemeProvider>
       </body>
     </html>
   );
