+++ frontend/components/ClientThemeProvider.tsx
@@ -1,0 +1,20 @@
+"use client";
+import React from "react";
+import { ThemeProvider, createTheme } from "@mui/material/styles";
+import CssBaseline from "@mui/material/CssBaseline";
+
+const theme = createTheme({
+  palette: {
+    primary: { main: "#3b82f6" },
+    secondary: { main: "#f97316" },
+  },
+});
+
+export default function ClientThemeProvider({ children }: { children: React.ReactNode }) {
+  return (
+    <ThemeProvider theme={theme}>
+      <CssBaseline />
+      {children}
+    </ThemeProvider>
+  );
+}
